refactor: split victory-config, personality, naming, and regiment catalogs (#4412)
Re-home leftover stall and declare-war param lists, and split personality types/tables/lookup, naming pools, and regiment rows so each library keeps one concern without changing public names or numeric defaults.
