+++ SPEC/program/repo-and-packages.md
@@ -203,7 +203,7 @@ Neither `colonizethis_logic` nor any logic-domain package (`colonizethis_world`,
 - **Given** the tech-ID convention gate reports a violation, **when** a developer inspects the output, **then** each violation includes file path, line, column, and the offending tech ID literal for direct remediation.
 - **Given** Dart source under `app/`, `packages/`, and `tool/`, **when** static analysis inspects executable AST string literals, **then** raw literals equal to canonical work target IDs are rejected outside the canonical work-target declaration file, whole-file scope skips for generated or catalog surfaces documented in `tool/check_work_target_constants.dart`, and approved fixture paths, with **no** keyed per-symbol waivers.
 - **Given** the work-target convention gate reports a violation, **when** a developer inspects the output, **then** each violation includes file path, line, column, and the offending work target literal with a suggested `kWorkTarget*` constant when available.
-- **Given** Dart source under `app/`, `packages/`, and `tool/`, **when** static analysis inspects executable AST string literals, **then** raw literals equal to canonical civilian unit type ids (`Explorer`, `Builder`, `Engineer`, `Spy`, `Merchant`, `Rail Builder` per `SPEC/game/civilian-units.md`) are rejected outside `packages/colonizethis_models/lib/src/civilian_unit_type_ids.dart`, approved fixture/test-data paths, and the single whole-file scope skip for `packages/colonizethis_data/lib/src/ai_personality_config.dart` (personality display strings that may collide with civilian spellings), with **no** keyed per-symbol waivers.
+- **Given** Dart source under `app/`, `packages/`, and `tool/`, **when** static analysis inspects executable AST string literals, **then** raw literals equal to canonical civilian unit type ids (`Explorer`, `Builder`, `Engineer`, `Spy`, `Merchant`, `Rail Builder` per `SPEC/game/civilian-units.md`) are rejected outside `packages/colonizethis_models/lib/src/civilian_unit_type_ids.dart`, approved fixture/test-data paths, and the single whole-file scope skip for `packages/colonizethis_data/lib/src/ai_personality_tables.dart` (personality display strings that may collide with civilian spellings), with **no** keyed per-symbol waivers.
 - **Given** the civilian unit type convention gate reports a violation, **when** a developer inspects the output, **then** each violation includes file path, line, column, and the offending literal with a suggested `kUnitType*` constant when available.
 
 ### Automated guard gate (CI)
