+++ packages/colonizethis_data/lib/src/combat_regiment_catalog_era1.dart
@@ -1,0 +1,75 @@
+import 'combat_config.dart';
+
+/// Era 1 regiment tactical rows. SPEC/game/military-units.md.
+const List<RegimentStats> regimentCatalogEra1 = [
+  RegimentStats(
+    id: 'peasant_levies',
+    fpn: 0,
+    fpm: 3,
+    rng: 1,
+    def: 3,
+    mvr: 3,
+    category: RegimentCategory.lightInfantry,
+    era: 1,
+  ),
+  RegimentStats(
+    id: 'pikemen',
+    fpn: 0,
+    fpm: 5,
+    rng: 1,
+    def: 5,
+    mvr: 3,
+    category: RegimentCategory.regularInfantry,
+    era: 1,
+  ),
+  RegimentStats(
+    id: 'arquebusiers',
+    fpn: 5,
+    fpm: 1,
+    rng: 3,
+    def: 3,
+    mvr: 2,
+    category: RegimentCategory.heavyInfantry,
+    era: 1,
+  ),
+  RegimentStats(
+    id: 'bowmen',
+    fpn: 3,
+    fpm: 1,
+    rng: 4,
+    def: 2,
+    mvr: 3,
+    category: RegimentCategory.bowmen,
+    era: 1,
+  ),
+  RegimentStats(
+    id: 'squires',
+    fpn: 0,
+    fpm: 4,
+    rng: 1,
+    def: 4,
+    mvr: 6,
+    category: RegimentCategory.lightCavalry,
+    era: 1,
+  ),
+  RegimentStats(
+    id: 'knights',
+    fpn: 0,
+    fpm: 6,
+    rng: 1,
+    def: 6,
+    mvr: 4,
+    category: RegimentCategory.spearCavalry,
+    era: 1,
+  ),
+  RegimentStats(
+    id: 'culverin',
+    fpn: 8,
+    fpm: 1,
+    rng: 5,
+    def: 2,
+    mvr: 2,
+    category: RegimentCategory.heavyArtillery,
+    era: 1,
+  ),
+];
