+++ packages/colonizethis_data/lib/src/combat_regiment_catalog_era2.dart
@@ -1,0 +1,85 @@
+import 'combat_config.dart';
+
+/// Era 2 regiment tactical rows. SPEC/game/military-units.md.
+const List<RegimentStats> regimentCatalogEra2 = [
+  RegimentStats(
+    id: 'calivermen',
+    fpn: 3,
+    fpm: 2,
+    rng: 5,
+    def: 5,
+    mvr: 4,
+    category: RegimentCategory.lightInfantry,
+    era: 2,
+  ),
+  RegimentStats(
+    id: 'halberdiers',
+    fpn: 0,
+    fpm: 7,
+    rng: 1,
+    def: 6,
+    mvr: 4,
+    category: RegimentCategory.regularInfantry,
+    era: 2,
+  ),
+  RegimentStats(
+    id: 'musketeers',
+    fpn: 7,
+    fpm: 2,
+    rng: 4,
+    def: 4,
+    mvr: 3,
+    category: RegimentCategory.heavyInfantry,
+    era: 2,
+  ),
+  RegimentStats(
+    id: 'cossacks',
+    fpn: 0,
+    fpm: 5,
+    rng: 1,
+    def: 5,
+    mvr: 8,
+    category: RegimentCategory.lightCavalry,
+    era: 2,
+  ),
+  RegimentStats(
+    id: 'lancers',
+    fpn: 0,
+    fpm: 8,
+    rng: 1,
+    def: 5,
+    mvr: 6,
+    category: RegimentCategory.spearCavalry,
+    era: 2,
+  ),
+  RegimentStats(
+    id: 'harquebusiers',
+    fpn: 2,
+    fpm: 6,
+    rng: 3,
+    def: 5,
+    mvr: 6,
+    category: RegimentCategory.heavyCavalry,
+    era: 2,
+  ),
+  RegimentStats(
+    id: 'horse_artillery',
+    fpn: 5,
+    fpm: 2,
+    rng: 7,
+    def: 2,
+    mvr: 3,
+    category: RegimentCategory.lightArtillery,
+    era: 2,
+  ),
+  RegimentStats(
+    id: 'royal_artillery',
+    fpn: 9,
+    fpm: 2,
+    rng: 8,
+    def: 2,
+    mvr: 2,
+    category: RegimentCategory.heavyArtillery,
+    era: 2,
+  ),
+];
