+++ packages/colonizethis_data/lib/src/combat_regiment_catalog_era3.dart
@@ -1,0 +1,75 @@
+import 'combat_config.dart';
+
+/// Era 3 regiment tactical rows. SPEC/game/military-units.md.
+const List<RegimentStats> regimentCatalogEra3 = [
+  RegimentStats(
+    id: 'skirmishers',
+    fpn: 4,
+    fpm: 3,
+    rng: 5,
+    def: 6,
+    mvr: 6,
+    category: RegimentCategory.lightInfantry,
+    era: 3,
+  ),
+  RegimentStats(
+    id: 'regulars',
+    fpn: 7,
+    fpm: 7,
+    rng: 5,
+    def: 5,
+    mvr: 4,
+    category: RegimentCategory.regularInfantry,
+    era: 3,
+  ),
+  RegimentStats(
+    id: 'grenadiers',
+    fpn: 10,
+    fpm: 8,
+    rng: 5,
+    def: 5,
+    mvr: 4,
+    category: RegimentCategory.heavyInfantry,
+    era: 3,
+  ),
+  RegimentStats(
+    id: 'hussars',
+    fpn: 2,
+    fpm: 8,
+    rng: 3,
+    def: 6,
+    mvr: 11,
+    category: RegimentCategory.lightCavalry,
+    era: 3,
+  ),
+  RegimentStats(
+    id: 'cuirassiers',
+    fpn: 5,
+    fpm: 13,
+    rng: 3,
+    def: 5,
+    mvr: 9,
+    category: RegimentCategory.heavyCavalry,
+    era: 3,
+  ),
+  RegimentStats(
+    id: 'light_artillery',
+    fpn: 8,
+    fpm: 3,
+    rng: 9,
+    def: 3,
+    mvr: 4,
+    category: RegimentCategory.lightArtillery,
+    era: 3,
+  ),
+  RegimentStats(
+    id: 'heavy_artillery',
+    fpn: 13,
+    fpm: 2,
+    rng: 10,
+    def: 2,
+    mvr: 3,
+    category: RegimentCategory.heavyArtillery,
+    era: 3,
+  ),
+];
