+++ packages/colonizethis_data/lib/src/combat_regiment_catalog_era4.dart
@@ -1,0 +1,75 @@
+import 'combat_config.dart';
+
+/// Era 4 regiment tactical rows. SPEC/game/military-units.md.
+const List<RegimentStats> regimentCatalogEra4 = [
+  RegimentStats(
+    id: 'sharpshooters',
+    fpn: 5,
+    fpm: 4,
+    rng: 7,
+    def: 7,
+    mvr: 7,
+    category: RegimentCategory.lightInfantry,
+    era: 4,
+  ),
+  RegimentStats(
+    id: 'rifle_infantry',
+    fpn: 9,
+    fpm: 9,
+    rng: 6,
+    def: 6,
+    mvr: 4,
+    category: RegimentCategory.regularInfantry,
+    era: 4,
+  ),
+  RegimentStats(
+    id: 'guards',
+    fpn: 12,
+    fpm: 10,
+    rng: 6,
+    def: 6,
+    mvr: 4,
+    category: RegimentCategory.heavyInfantry,
+    era: 4,
+  ),
+  RegimentStats(
+    id: 'scouts',
+    fpn: 5,
+    fpm: 11,
+    rng: 5,
+    def: 6,
+    mvr: 11,
+    category: RegimentCategory.lightCavalry,
+    era: 4,
+  ),
+  RegimentStats(
+    id: 'carbine_cavalry',
+    fpn: 7,
+    fpm: 17,
+    rng: 5,
+    def: 5,
+    mvr: 9,
+    category: RegimentCategory.heavyCavalry,
+    era: 4,
+  ),
+  RegimentStats(
+    id: 'field_artillery',
+    fpn: 10,
+    fpm: 3,
+    rng: 11,
+    def: 4,
+    mvr: 5,
+    category: RegimentCategory.lightArtillery,
+    era: 4,
+  ),
+  RegimentStats(
+    id: 'siege_guns',
+    fpn: 17,
+    fpm: 2,
+    rng: 12,
+    def: 3,
+    mvr: 3,
+    category: RegimentCategory.heavyArtillery,
+    era: 4,
+  ),
+];
