+++ packages/colonizethis_data/lib/src/naming_default_great_powers.dart
@@ -1,0 +1,189 @@
+import 'naming_rules.dart';
+
+/// Great Power naming pools (GDD 09).
+const List<GreatPowerNaming> defaultNamingGreatPowers = [
+  GreatPowerNaming(
+    id: 'england',
+    countryName: 'England',
+    adjective: 'English',
+    capitalCityName: 'London',
+    leaderVariants: [
+      LeaderVariant(
+        id: 'queen_victoria',
+        name: 'Queen Victoria',
+        leaderKey: 'england_leader',
+        provinceNamePool: [
+          'Kent',
+          'Yorkshire',
+          'Lancashire',
+          'Cornwall',
+          'East Anglia',
+          'Wessex',
+          'Northumberland',
+          'Gloucestershire',
+          'Lincolnshire',
+        ],
+      ),
+    ],
+  ),
+  GreatPowerNaming(
+    id: 'france',
+    countryName: 'France',
+    adjective: 'French',
+    capitalCityName: 'Paris',
+    leaderVariants: [
+      LeaderVariant(
+        id: 'napoleon',
+        name: 'Napoleon Bonaparte',
+        leaderKey: 'france_leader',
+        provinceNamePool: [
+          'Normandy',
+          'Brittany',
+          'Guyenne',
+          'Languedoc',
+          'Provence',
+          'Burgundy',
+          'Champagne',
+          'Picardy',
+          'Dauphiné',
+        ],
+      ),
+    ],
+  ),
+  GreatPowerNaming(
+    id: 'spain',
+    countryName: 'Spain',
+    adjective: 'Spanish',
+    capitalCityName: 'Madrid',
+    leaderVariants: [
+      LeaderVariant(
+        id: 'isabella',
+        name: 'Queen Isabella I',
+        leaderKey: 'spain_leader',
+        provinceNamePool: [
+          'Old Castile',
+          'Aragon',
+          'Catalonia',
+          'Andalusia',
+          'Valencia',
+          'Galicia',
+          'Navarre',
+          'Granada',
+          'León',
+        ],
+      ),
+    ],
+  ),
+  GreatPowerNaming(
+    id: 'portugal',
+    countryName: 'Portugal',
+    adjective: 'Portuguese',
+    capitalCityName: 'Lisbon',
+    leaderVariants: [
+      LeaderVariant(
+        id: 'henry_navigator',
+        name: 'Prince Henry the Navigator',
+        leaderKey: 'portugal_leader',
+        provinceNamePool: [
+          'Beira',
+          'Alentejo',
+          'Algarve',
+          'Minho',
+          'Trás-os-Montes',
+          'Douro Litoral',
+          'Ribatejo',
+          'Beira Baixa',
+          'Estremadura do Norte',
+        ],
+      ),
+    ],
+  ),
+  GreatPowerNaming(
+    id: 'netherlands',
+    countryName: 'Netherlands',
+    adjective: 'Dutch',
+    capitalCityName: 'Amsterdam',
+    leaderVariants: [
+      LeaderVariant(
+        id: 'de_ruyter',
+        name: 'Michiel de Ruyter',
+        leaderKey: 'netherlands_leader',
+        provinceNamePool: [
+          'Zeeland',
+          'Utrecht',
+          'Friesland',
+          'Groningen',
+          'Overijssel',
+          'Gelderland',
+          'Drenthe',
+          'Flanders',
+          'Brabant',
+        ],
+      ),
+    ],
+  ),
+  GreatPowerNaming(
+    id: 'prussia',
+    countryName: 'Prussia',
+    adjective: 'Prussian',
+    capitalCityName: 'Berlin',
+    leaderVariants: [
+      LeaderVariant(
+        id: prussiaVariantFrederickTheGreat,
+        name: 'Frederick the Great',
+        leaderKey: 'prussia_leader',
+        provinceNamePool: [
+          'Pomerania',
+          'Silesia',
+          'East Prussia',
+          'Magdeburg',
+          'Halberstadt',
+          'Minden',
+          'Ravensberg',
+          'Neumark',
+          'Jülich-Cleves',
+        ],
+      ),
+      LeaderVariant(
+        id: prussiaVariantFrederickWilliam,
+        name: 'Frederick William',
+        leaderKey: 'prussia_reserve_leader',
+        provinceNamePool: [
+          'Prussia',
+          'Pomerania',
+          'Farther Pomerania',
+          'Magdeburg',
+          'Halberstadt',
+          'Minden',
+          'Ravensberg',
+          'Neumark',
+          'Jülich-Cleves',
+        ],
+      ),
+    ],
+  ),
+  GreatPowerNaming(
+    id: 'sweden',
+    countryName: 'Sweden',
+    adjective: 'Swedish',
+    capitalCityName: 'Stockholm',
+    leaderVariants: [
+      LeaderVariant(
+        id: 'gustavus',
+        name: 'Gustavus Adolphus',
+        leaderKey: 'sweden_leader',
+        provinceNamePool: [
+          'Västmanland',
+          'Skåne',
+          'Östergötland',
+          'Västergötland',
+          'Småland',
+          'Dalarna',
+          'Norrland',
+          'Finland',
+          'Estonia',
+        ],
+      ),
+    ],
+  ),
+];
